requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bandit==1.8.3
certifi==2025.1.31
click==8.1.8
fastapi==0.115.12 # Se Mantiene versión actual (sin vulnerabilidades conocidas)
h11==0.16.0 # Se pasa de 0.14.0 a 0.16.0 para corregir/remediar CVE-2025-43859 (HTTP Request Smuggling)
httpcore==1.0.9 # Se pasa de 1.0.7 a 1.0.9 para corregir/remediar CVE-2025-43859 (HTTP Request Smuggling)
flake8==7.1.2
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
packaging==24.2
pbr==6.1.1
pluggy==1.5.0
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1 # Versión segura (mantener)
pytest==8.3.5
python-dotenv==1.0.1
PyYAML==6.0.2
rich==13.9.4
setuptools==78.0.2
sniffio==1.3.1
starlette==0.46.1
stevedore==5.4.1
typing_extensions==4.12.2
uvicorn==0.34.0 # Versión segura (mantener)
gunicorn==23.0.0 # Se pasa de 21.2.0 a 23.0.0 para corregir/remediar CVE-2024-1135 y CVE-2024-6827 (HTTP Smuggling)



